+++ src/abx_amr_simulator/training/__init__.py
@@ -1,5 +1,9 @@
 """Training utilities and entrypoints."""
 
 from .setup_tuning import setup_optimization_folders_with_defaults
+from .tune_marl_agents import tune_marl_agents_sequentially
 
-__all__ = ['setup_optimization_folders_with_defaults']
+__all__ = [
+    'setup_optimization_folders_with_defaults',
+    'tune_marl_agents_sequentially',
+]